Move app into seperate folder
