+++ src/App/App.js
@@ -5,13 +5,13 @@ import {
   getMovieDataAPI,
   getMovieDetailsAPI,
   getMovieVideoAPI,
-} from "./components/utilities/apiCalls";
-import {checkBudgetInfo} from './components/utilities/dataCleaning'
-import ErrorPage from "./components/errorPages/ErrorPage";
-import MovieDetails from "./components/MovieDetails/MovieDetails";
-import NavBar from "./components/NavBar/NavBar";
-import Movies from "./components/Movies/Movies";
-import SearchResults from "./components/SearchResults/SearchResults"
+} from "../components/utilities/apiCalls";
+import {checkBudgetInfo} from '../components/utilities/dataCleaning'
+import ErrorPage from "../components/errorPages/ErrorPage";
+import MovieDetails from "../components/MovieDetails/MovieDetails";
+import NavBar from "../components/NavBar/NavBar";
+import Movies from "../components/Movies/Movies";
+import SearchResults from "../components/SearchResults/SearchResults"
 
 class App extends Component {
   state = {
